change: tms: move pakuru.ts to tms/ dir
